+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 - **fix**: Constantly check CPH status
 - **fix**: Remove unsupported schema keyword minimum
 - **feat**: Add unzip folder configuration
+- **feat**: Add placeholder `basenameNoExt`
 
 ## 0.0.8